Correct angles and energy functions
Fixed the problem with the NaN positions popping out, generated by situations in which the domain of the acos function was not respected (sometimes the argument would get slightly bigger than one, most likely because of errors in approximation of the distance); still not sure if there are errors in the computation of the bead-bead distance.

Added two functions computing kinetic and potential energy; it turns out that the total energy is not perfectly conserved in the case in which I consider only elastic forces. This probably hides some error in the code.